+++ src/main/java/com/sparta/calendarprojectsnext/domain/comment/command/CommentCommand.java
@@ -7,7 +7,7 @@ import com.sparta.calendarprojectsnext.domain.exception.CustomException;
 import com.sparta.calendarprojectsnext.domain.schedule.entity.Schedule;
 import com.sparta.calendarprojectsnext.domain.schedule.repository.ScheduleRepository;
 
-import static com.sparta.calendarprojectsnext.domain.exception.ErrorCode.SCHEDULE_NOT_FOUND;
+import static com.sparta.calendarprojectsnext.domain.exception.eunm.ErrorCode.SCHEDULE_NOT_FOUND;
 
 public class CommentCommand {
     public static class Create {
