+++ src/main/java/com/sparta/calendarprojectsnext/domain/exception/CustomException.java
@@ -1,5 +1,6 @@
 package com.sparta.calendarprojectsnext.domain.exception;
 
+import com.sparta.calendarprojectsnext.domain.exception.eunm.ErrorCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
