+++ src/main/java/com/sparta/calendarprojectsnext/domain/schedule/command/ScheduleCommand.java
@@ -10,7 +10,7 @@ import com.sparta.calendarprojectsnext.domain.userschedule.entity.UserSchedule;
 
 import java.time.LocalDateTime;
 
-import static com.sparta.calendarprojectsnext.domain.exception.ErrorCode.USER_NOT_FOUND;
+import static com.sparta.calendarprojectsnext.domain.exception.eunm.ErrorCode.USER_NOT_FOUND;
 
 public class ScheduleCommand {
